requirements.txt
streamlit
yfinance
pandas
pandas_ta
plotly
requests
google-generativeai
feedparser